Fix pin setup to camera navigation
There is still a timing issue in the handlePINSetupComplete and I had to resort to setting self?.hasCompletedIntro = true to get this navigation to function. We might need to dig further on that timing issue. It comes down to when these "has completed setup" and "is authed" values are set vs when the observers see them and navigate.