Fix race condition on camera permission granted
There was a race when updating the camera permission status. We initialize at false. The user responds by granting access. When we come back to the view, the shutter button and view being shown are still treating the user like they didn't just authorize camera access. That's stupid.

Observe the state for changes to allow the view to update.

Add a check for in-progress permissions changes and allow them to happen.