Improve photo to gallery transition
This solves a small bug where taking a pic and clicking the gallery immediately allows the user to view the gallery, but the images aren't there yet. We are still processing the image. We could disable the button with no feedback but it's better to disable it with a small spinner to indicate progress.