Prevent showing last camera image on start
This fixes a camera view/visibility issue. When the app is in the camera view and sent to background, the current image on the screen at the time the app minimized is frozen in the view. Bringing the app back to the foreground will flash this previous image for a moment until the new camera image appears on the screen. This was bad UX.

Throw a black overlay in and give a bit of time for the camera and its image to start back up.