Refactor photo detail view
This is a large change to get the photo detail view to look and feel like all the other gallery apps on ios. The previous implementation was obnoxious because it couldn't discern between panning and zooming. We can't use the tabview because the swipe gestures were interfering with the gestures we use to zoom/pan images.

- pinch to zoom
- single and double finger pan

Then we have to handle the actual swiping right/left to adjacent images. We also have to handle downward swipes to dismiss, but we're only going to do that when the image is zoomed out.

One other option to make it feel more natural would be to dump that "info" button in the toolbar, then make an image swipe upward expose the info sheet below. That's not changed here though. Just an idea.