Add info sheet into the main nav stack
This page wasn't inside the new nav stack we're making here. Adding it in now.

The reason we have to use .securitymanaged() twice more here is sheets have their own view hierarchy so we must be setting this at their root. Everything's in ContentView now so we have one central place/view to manage where the security overlay is called.